requirements.txt
ahk==0.13.0
aiofiles==0.8.0
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.7.7
alexapy==1.25.3
ansi2html==1.6.0
anyio==3.4.0
APScheduler==3.6.3
async-timeout==4.0.2
attrs==21.4.0
authcaptureproxy==1.1.1
backoff==1.11.1
beautifulsoup4==4.10.0
bidict==0.22.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cryptography==35.0.0
dateparser==1.1.1
distlib==0.3.4
dnspython==1.16.0
et-xmlfile==1.1.0
eventlet==0.30.0
filelock==3.6.0
Flask==1.0.2
Flask-Cors==3.0.10
Flask-Excel==0.0.7
Flask-Login==0.4.1
Flask-Migrate==2.6.0
Flask-Script==2.0.6
Flask-Session==0.3.1
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.5.1
frozenlist==1.3.0
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
httpcore==0.14.3
httpx==0.21.1
idna==3.3
importlib-metadata==4.11.3
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.10
lml==0.1.0
Mako==1.2.0
MarkupSafe==1.1.0
multidict==5.2.0
numpy==1.22.3
openpyexcel==2.5.14
openpyxl==3.0.9
pandas==1.4.2
platformdirs==2.5.2
protobuf==3.19.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
pyexcel-xlsx==0.6.0
pyotp==2.6.0
python-binance==1.0.15
python-dateutil==2.8.2
python-dotenv==0.20.0
python-engineio==3.13.2
python-socketio==4.6.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.2
requests==2.27.1
rfc3986==1.5.0
simplejson==3.17.6
six==1.11.0
sniffio==1.2.0
soupsieve==2.3.1
SQLAlchemy==1.4.36
texttable==1.6.4
typer==0.4.0
typing-extensions==4.0.1
tzdata==2022.1
tzlocal==4.2
ujson==5.2.0
Unidecode==1.3.4
urllib3==1.26.9
virtualenv==20.14.1
websocket-client==0.57.0
websockets==9.1
Werkzeug==0.14.1
wsproto==1.1.0
xlrd==2.0.1
yarl==1.7.2
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
